requirements.txt
openai
azure-ai-evaluation
python-dotenv
bs4
azure-identity
pandas
qrcode[pil]
azure-search-documents==11.6.0b4
streamlit
PyPDF2
azure-ai-projects
azure-monitor-opentelemetry
ag2
pyautogen
azure-ai-inference
opentelemetry-instrumentation-openai-v2
semantic-kernel
autogen-core
autogen-agentchat
autogen-ext[openai]
autogen-ext[magentic-one,openai]
azure-core-tracing-opentelemetry
pyodbc
autogen-ext
python-docx
pymupdf